Fix size of log module in AppSettings. A hung socket is not necessarily a hack attemp

+++ libs/CoreLib/VxDebug.cpp
@@ -39,7 +39,7 @@ bool g_StreamActive = false;
 namespace
 {
 #if defined(LOG_IN_RELEASE_BUILD)
-uint64_t g_ModuleEnableLoggingFlags = (uint32_t)(
+uint64_t g_ModuleEnableLoggingFlags = (uint64_t)(
     eLogHackers
     // | eLogFileXfer
     // | eLogConnect
@@ -72,7 +72,7 @@ uint64_t g_ModuleEnableLoggingFlags = (uint32_t)(
     //| eLogVideoIo
     );
 #elif defined(CAMERATESTBUILD)
-        uint64_t g_ModuleEnableLoggingFlags = (uint32_t)(
+        uint64_t g_ModuleEnableLoggingFlags = (uint64_t)(
         eLogHackers
         // | eLogFileXfer
         // | eLogConnect
@@ -106,7 +106,7 @@ uint64_t g_ModuleEnableLoggingFlags = (uint32_t)(
         );
 
 #elif defined(DEBUG) || defined(_DEBUG) || defined(FLATPAKBUILD)
-    uint64_t g_ModuleEnableLoggingFlags = (uint32_t)(
+    uint64_t g_ModuleEnableLoggingFlags = (uint64_t)(
         eLogHackers
         // | eLogFileXfer
         // | eLogConnect
